Added character direction always forward and an interactable script
The character now turns to the direction they are walking, and there is now a interactable rectangle that turns green if you press space around it.

I am very tired right now. There was a bug with the new movement where it kept going after the button was pressed. I found the solution two hours ago. I did not realize that I had fixed it till 15 mins ago. I am happy that it works. That is all I have to say.