Dynamic camera movement
We want some scenes/areas to be long/tall, so I updated our PlayerMovement script to include camera movement alongside the player's. The camera is clamped so it does not go past the boundaries of the screen, and the player remains centered if they turn back from being past the camera boundaries so that the design can remain consistent. I have also started on changing these boundaries regarding what plane (scene/area) the player is on. This will be expanded in the future once I get test it more accurately following scene/area transitions.