Camera shots between planes/areas
We are going to have different planes/areas that the player will traverse within a single town, so I created a test for this. It involves trigger points and empty objects representing camera and player shots, which practically just teleports both onto the correlated empty object when the player collides with the correlated trigger point.