Plane # Changes Per Shot
As previously referenced, this is an expanded tracking method for tracking plane numbers, so the PlayerMovement script can track what plane/area the player is currently on. This plane number will then be used to specify the camera's boundaries for that specific plane/area, which is already used for the 1st plane/area.