Added initial overworld
We wanted a curved world that hides locales from the player, nudging them to explore the overworld. I found this shader that mimics Animal Crossing's curved world and used it for our overworld: https://alastaira.wordpress.com/2013/10/25/animal-crossing-curved-world-shader/

The shader curves anything that is attached to it along the camera. I also updated player controls specifically for the overworld and made them feel like Final Fantasy 9 tank controls to simulate that game's traversal.