Added player shadow in overworld
The traditional casted shadow provided by Unity did not translate well with the curved world. Because the curved world is not super smooth it looks like the player is making super long jumps with casted shadow. I made a drop shadow so the shadow stays with the player at all times.