Overworld: forest and waterfall
I added a forest and waterfall to the overworld among some smaller things and updates. All of this was added to give the world a stronger environment for the player to latch onto. Every fantasy world needs a forest of course, and the waterfall gives the second town some dynamic imagery.

The forest has a long forest wall to block off the northmost boundary. This wall uses Ocarina of Time wall boundary textures. I added an entrance and stairs down to the dock, all with ps1 style textures. Finally, the waterfall and new river uses the same ps1-esque flowing water material. All of these were used to give our game an old ps1/n64 feeling.