Added camp and forest hideaway
I added a camp and kid's hideaway to the forest. All of what I added and how it looks is based on the descriptions I was given or what assets I could find that worked. For the camp, I wanted a semi-permanent but easy and simply looking camp, so it is pretty much just a bed and campfire but also has some barrels to simulate storage. I placed it by the river because maybe that would be a source of water and to fill out that side of the map a bit. For the forest hideaway, I was told it is like a kid's hideaway that they made in the forest, so I placed it a little further back in the forest to try and hide it a bit but still close to the edge since kids would probably not want to be too far out. All it is is a tent and some log seats, and thankfully (it took me forever) I found a good-looking tent model with a patchy texture, so it kind of looks like a kid snatched some spears and stitched together some probably also stolen clothes.