Added props to forest hideaway
Tasked with adding sticks, rocks, a campfire, and a wooden sword, shield, and slingshot "for the young boy melio's arc." I placed these items in a scattered about location to hopefully get the design that it is a kid's hideaway. All of the assets were found with a PS1/N64/lowpoly style in mind like most everything else to, again, simulate that PS1 era of fantasy world.