Melio overworld movement
This commit adds in AI Navigation for NPC movement (specifically for the overworld, but its scalable I hope) alongside a clock system.

I baked a NavMesh into the overworld terrain so I could have the NPC Melio traverse from town A to his forest hideaway seamlessly and without me having to hard code any navigations. I tried using Vector3.MoveTowards but that is very braindead as the NPC does not have any AI to register its surroundings; Melio just kept running into things. I then was about to create an animation before doing AI navigation, but I realized that because of all of the obstacles in the overworld (notably the painted trees), it would take me a long while to make an animation that looked good. I created a Melio-specific and an NPCNavigator script. The NPCNavigator script simply gets the assigned object's NavMeshAgent component and houses a Move() function that sets the object's destination to a determined target. I did not want to put anything too specific into this file so we can use it across multiple NPCs, like I do in the Melio script. The Melio script determines what scene it is currently in, what time it is, and what it should do if those parameters line up. For example, I have it checking: if it is in the overworld (the first scene) and if it is 1:30pm/13:30, then call the NPCNavigator's Move() function to go to the hideaway. This is the foundation to have our NPCs navigate throughout our scenes to simulate their realism and have them at their designated locations at the correct times.

The clock system is all determined by the new GameManager script. This script has the DoNotDestroyOnLoad() function called in it so hopefully it should remain a single instance across any scene we traverse as long as it is created in our introductory scene, which is the overworld. This will allow a clock to flow linearly throughout the entire game, allowing us to schedule events during certain times like what I did with Melio, and hopefully have them happen across scenes. All of this is to simulate some realism and passage of time, a key part to the message of our game and how the player/character should value their time. I created a simple UI element for the clock to, mainly for testing but we may keep it in so the player can keep track

This is a foundational commit. I am sure much of this will improve over time.