Added Bennet's character expressions
Bennet is a grumpy old farmer, so I wanted his character design to portray that. He is a stout man with a frown on his face, and his color palette is a dusty brown like the soil he tends to. I've only given the non-important characters, like Bennet, Sybil, and Abigail, 2 neutral faces to exchange from, since they don't have as much dialouge.