Created Abigail's expressions
Abigail is the young tavern-keeper in the town, with bright red hair, freckles, and a big grin. She has a crush on the boy in the town over, and Irwin was supposed to help those two get together, but that didn't end up happening. I actually had her hair much redder, but I felt like the brightness didn't go well enough with the overall color scheme of the game.